Project - Improving 404 error handler

+++ forum/src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -63,23 +64,34 @@ public class TopicosController {
 	}
 	
 	@GetMapping("/{id}")
-	public TopicoPorIdDto listarPorId(@PathVariable Long id) {
-		Topico topico = topicoRepository.getById(id);
-		return new TopicoPorIdDto(topico);
+	public ResponseEntity<TopicoPorIdDto> listarPorId(@PathVariable Long id) {
+		//Optional<Topico> topico = topicoRepository.findById(id);
+		return topicoRepository.findById(id)
+                .map(TopicoPorIdDto::new)
+                .map(ResponseEntity::ok)
+                .orElse(ResponseEntity.notFound().build());
 	}
 	
 	@Transactional
 	@PutMapping("/{id}")
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid TopicoFormAtalizacao form) {
-		Topico topico = form.atualizar(id, topicoRepository);
-		return ResponseEntity.ok(new TopicoDto(topico));
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if (optional.isPresent()) {
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}
+		return ResponseEntity.notFound().build();
 	}
 	
 	@Transactional
 	@DeleteMapping("/{id}")
 	public ResponseEntity<?> excluir(@PathVariable Long id) {
-		topicoRepository.deleteById(id);
-		return ResponseEntity.noContent().build();
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if (optional.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.noContent().build();
+		}
+		return ResponseEntity.notFound().build();
 	}
 	
 }
